rename package to conventional name
