--- src/main/java/org/dredd/bulletCore/models/ammo/Ammo.java
@@ -1,24 +1,0 @@
-package org.dredd.bulletCore.models.ammo;
-
-import org.bukkit.inventory.ItemStack;
-import org.dredd.bulletCore.models.CustomBase;
-import org.jetbrains.annotations.NotNull;
-
-/**
- * Represents ammo items.
- *
- * @author dredd
- * @since 1.0.0
- */
-public class Ammo extends CustomBase {
-
-    public Ammo(BaseAttributes attrs) {
-        super(attrs);
-    }
-
-    @Override
-    public @NotNull ItemStack createItemStack() {
-        // Add only ammo-specific attributes
-        return createBaseItemStack();
-    }
-}
